requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.8.0
pydantic-settings>=2.2.1
opencv-python         
numpy>=2.1.0
requests==2.31.0
aiofiles==23.2.1
#mediapipe==0.10.1
pillow
